routes: extract rules + nat (#187 second slice) (v5.95.2)
Continuing the routes/mod.rs split started in #266. This commit
extracts the two largest CRUD resources — firewall rules and NAT
rules — into their own modules.

What landed:

- `routes/rules.rs` (342 lines) — `CreateRuleRequest`, `ReorderRequest`,
  + 8 handlers: list_system_rules, list_rules, get_rule, create_rule,
  update_rule, delete_rule, toggle_block_logging, reorder_rules.
- `routes/nat.rs` (203 lines) — `CreateNatRuleRequest` + 6 handlers:
  list_nat_rules, create_nat_rule, update_nat_rule, delete_nat_rule,
  get_nat_pf_output, reorder_nat_rules. Imports `ReorderRequest` from
  `super::rules` since both endpoints share the same payload shape.

mod.rs is now 3160 lines (down from 3689 after slice 1, down from
4013 originally — total reduction so far: 853 lines / 21%).

Resources still in mod.rs awaiting extraction (each can land in a
follow-up commit with the same pattern):
  auth (~15 handlers), users + roles (~13), vpn (~17), backup (~2),
  static_routes (~4), interfaces (~2), settings (~14),
  status/connections/logs/metrics (~9).

Verified:
- `cargo check --workspace` clean
- `cargo test --package aifw-api` — 123 tests pass
