+++ aifw-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aifw-ui",
-  "version": "5.95.1",
+  "version": "5.95.2",
   "private": true,
   "scripts": {
     "dev": "next dev",
